Dependencies cleanup Restructuring
Signed-off by: Sorin Moldovan <sorin@hpe.com>
