+++ cs-content-generator-sources/src/main/java/io/cloudslang/tools/generator/CsGenerator.java
@@ -9,10 +9,10 @@ import com.hp.oo.sdk.content.annotations.Action;
 import com.hp.oo.sdk.content.annotations.Output;
 import com.hp.oo.sdk.content.annotations.Param;
 import com.hp.oo.sdk.content.annotations.Response;
-import io.cloudslang.tools.generator.entities.cs.CsOperationFile;
+import io.cloudslang.tools.generator.entities.CsOperationFile;
 import io.cloudslang.tools.generator.services.OperationService;
-import io.cloudslang.tools.generator.services.fs.JarService;
-import io.cloudslang.tools.generator.services.fs.MavenService;
+import io.cloudslang.tools.generator.services.JarService;
+import io.cloudslang.tools.generator.services.MavenService;
 import javassist.ClassClassPath;
 import javassist.ClassPool;
 import javassist.CtClass;
