+++ cs-content-generator-sources/src/main/java/io/cloudslang/tools/generator/entities/Mappable.java
@@ -1,0 +1,7 @@
+package io.cloudslang.tools.generator.entities;
+
+import java.util.Map;
+
+public interface Mappable {
+    Map<String, Object> toMap();
+}
