--- cs-content-generator-sources/src/main/java/io/cloudslang/tools/generator/entities/cs/Mappable.java
@@ -1,10 +1,0 @@
-package io.cloudslang.tools.generator.entities.cs;
-
-import java.util.Map;
-
-/**
- * Created by moldovso on 5/24/2017.
- */
-public interface Mappable {
-    Map<String, Object> toMap();
-}
