+++ cs-content-generator-sources/src/main/java/io/cloudslang/tools/generator/services/converters/ResultExpressionConverterService.java
@@ -2,10 +2,6 @@ package io.cloudslang.tools.generator.services.converters;
 
 import com.hp.oo.sdk.content.plugin.ActionMetadata.MatchType;
 
-/**
- * Author: Ligia Centea
- * Date: 4/26/2016.
- */
 public class ResultExpressionConverterService {
 
     private static final String ARGUMENT = "%s";
@@ -14,7 +10,7 @@ public class ResultExpressionConverterService {
     public static String getMatchingExpression(MatchType matchType) {
         switch (matchType) {
 
-            //compare arguments
+            //Compare arguments
             case COMPARE_NOT_EQUAL:
                 return ARGUMENT + "!=" + QUOTED_ARGUMENT;
             case COMPARE_LESS:
@@ -28,14 +24,13 @@ public class ResultExpressionConverterService {
             case COMPARE_GREATER_OR_EQUAL:
                 return ARGUMENT + ">=" + QUOTED_ARGUMENT;
 
-            //containment arguments
+            //Containment arguments
             case CONTAINS:
                 return ARGUMENT + " in " + ARGUMENT;
             case EXACT:
                 return ARGUMENT + "==" + QUOTED_ARGUMENT;
             case NOT_EXACT:
                 return ARGUMENT + "!=" + QUOTED_ARGUMENT;
-            //TODO -> missing items here
             case BEGINS_WITH:
                 return ARGUMENT + ".startswith(" + QUOTED_ARGUMENT + ")";
             case ENDS_WITH:
