+++ cs-content-generator-sources/src/main/resources/templates/CloudSlangOperation.hbs
@@ -1,6 +1,6 @@
 ########################################################################################################################
 #!!
-#! @description: Generated operation description
+#! @description: Generated operation description.
 #!
 {{{documentation}}}
 #!!#
